"2:15pm. I am tired. Honestly, if every day gave me fortuitous encounters like this, I'd long have achieved transcendence.
Yesterday I gave up on the Blender plugin plan, but now it occurs to me that it is alive again. Which is just as well, since my plan was to study the code, but now I have an actual reason for doing that.

2:25pm. When it is time to resume, I think what I will do is first post this batch on my Twitter.

3:10pm. Enough Legendary Mechanic. It is time to start getting ready for the next session. Let chill a bit more and I will start.

3:55pm. I've done a few more images. Right now I am forming an opinion on CAST. It is pretty good. Compared to other methods it is less stylized and preserves the salient features of the original better. But whether it is better depends on what you are going for. Still, I will go for it as it is so easy to use compared to NNST.

I've been trying to make ML work for me for so long, and look at this. It came to me in literally a snap. Without a doubt, this path is the true path. Being able to do style transfer like this eliminates the primary weakness of 3d.

5pm. Done with the tweeting spree.

5:20pm. Somehow I really like that image of the crab I posed in the wip thread. It is not noisy in the way that my own image is. In fact it is clean in the way I'd prefer the anime style to be. If I want a cleaner look I'll really have to learn how to denoise my Luxcore products. I think I turned on the option, so why didn't it do the job?

No, actually I did not turn on the option. The reason for that is because it is not using Nvidia's denoiser, but the slow CPU ones. I didn't want to bother testing if it does it during rendering or just at the end. If I ever use Luxcore again I should make it a priority to try it out.

5:25pm. Looking at the crab image it occured to me that the nature of the blur operation changes with CAST. Rather than the image really being blurry like in the photo, it is instead drawn with large strokes. Contrasted with the high detail, tiny strokes used to draw the drab, it looks really great. It has a ton of watermarks over it, which I did not post on Twitter.

5:30pm. I am just engrossed in my own imagination. I need to get on with it.

Even though I've been doing nothing, but making posts all day today, it took a lot out of me. I've been mentally active for 9h already.

5:40pm. https://github.com/junyanz/pytorch-CycleGAN-and-pix2pix/blob/master/docs/qa.md

Right now I am reading this. It seems the CAST authors are either related to the CycleGAN authors or they've adapted their code in a big way.

5:45pm. I am thinking what I want to do right now. I do not have the mental energy for an in depth code dive. Let me read the paper again. What did they say about future improvements?

> In the future, we plan to improve the contrastive style learning process by considering artist and category information

Ah, I see.

5:50pm. Right now, I do not really understand how CAST works. Vanilla style transfer sure, but CAST is doing something else.

6:20pm. > To maintain the content information of the content image in the process of style transfer between the two domains, we also add a cycle consistency loss:

What is this for? How does this make sense. I think the repo does not have the training code yet.

> The results of using other backbones are shown in the supplementary materials.

Supplementary materials?

6:30pm. I've read the paper again. I think I understand it better now, but I do not really understand the paired inputs in the generator. I'll want to check out Cycle GAN and some of the references they are borrowing their tricks from.

6:35pm. Hmmmm, no let's leave it at this. My brain needs time to unwind.

Come to think of it, NN denoising is a thing. Since it is too late to do it the regular way, maybe I should try that?

Nevermind it for now. My goal for tomorrow should be simple - I'll walk myself through the code and see if I can go even higher than 1.4k. I can't imagine not having neough memory for 2k images. It is one thing to hold a large number of intermediaries through 30 layers of training, but passing it feedforward through the net should not be a problem even for my GTX 970.

6:40pm. In order to achieve that goal, I will need more insight into the code, hence I have a motivation to actually study it besides pure curiosity.

Let me have lunch. I am done for the day."