"10:10am. Let me finish my morning reading and then I will get started.
https://www.bilibilicomics.com/mc480/175541?from=manga_detail

The coloring in Egg of the Elf is something that grows on me the more I read it.

10:35am. Let me start. I've slacked enough.

https://news.ycombinator.com/item?id=31699468
AI Could Defeat All of Us Combined

///

What a limited imagination of how an AI would take over. The scenarios seem to be centered around an extrapolation of "What if a really smart human were trapped inside a computer?" The amount of anthropomorphism is astounding.
The author misses an even scarier prospect - people will want to run such an AI. They will be absolutely giddy at the prospect of running such an AI and it won't be anything like a really smart human trapped in a computer.

AI is already laying the groundwork if you look around today. Every other tweet is a DALL-E[1] image. They are everywhere. DALL-E is increasing its reach while simultaneously signaling that it is an area of research worth pursuing. In effect kicking off the next generation of image generating AIs.

Generation is an apt term. We can utilize the language of organisms with ease. DALL-E lives by way of people invoking it, and reproduces by electro-memeticly - someone else viewing the output and deciding to run DALL-E themselves. It undergoes variation and selection. As new research takes place, and produces new models, they succeed by producing images which further its reproduction, or it doesn't and the model is an evolutionary dead-end.

AI physiologically lives on the cost to run it, and evolves at the rate of research applied. Computational reserves and mindshare are presently fertile new expanses for AI, but what occurs when resources are constrained and inter-AI conflict rises? I expect the result to look similar to competition between parasites for a host - a complex multi-way battle for existence. But no, nothing like a deranged dictator scenario. Leave that for the movies.

1. or variant thereof

///

Yesterday I found this comment to be pretty good so let me highlight it. The article itself does not seem to be worth reading.

I am amazed at how much /ic/ seems to be losing its shit over it.

10:40am. Let me get started. Let me resume the tutorial from yesterday.

https://youtu.be/9Q8PwcDzb8Y?t=2056
How to Make Earth in Blender (Cycles)

https://youtu.be/9Q8PwcDzb8Y?t=2010

Hmmm, I am confused how he gets such diffuse lighting from the sun's reflection.

...Yeah, it does not make sense for the ocean to be a perfect mirror like it is now. At this scale, the waves should act as micro facets. I've cranked up the roughness using a color ramp and now it seems better.

https://stock.adobe.com/search?k=deep+space+hdri

Adobe has a lot HDRIs.

https://www.youtube.com/watch?v=sksfsOSlGkQ
6 FREE SPACE HDRI ENVIRONMENT MAPS

Only one of the images in here would be good for my purpose. The rest have planets and objects in them.

11am. Ah, goddamit. let me import the one I made months ago myself.

https://www.reddit.com/r/blender/comments/hdg44i/360_actual_night_sky_hdri_using_nasa_data/

I want to get NASA's maps, but it site seems to be down.

https://stock.adobe.com/search?k=deep+space+hdri&asset_id=266760651

> Must not be used for merchandise for sale. How annoying.

https://nasa3d.arc.nasa.gov/

Finally, I accessed the NASA site!

11:20am. Just where did andrew find those maps of the Earth? He says it is from NASA, but I can't find it. It is not that easy to find resources. Various sites have stock photos of deep space, and they require a subscription, but I am having surprising trouble just finding something I could get directly. I really will end up using the limbo asset.

https://www.youtube.com/watch?v=GG7aydoWxb0
JOURNEY TO GALAXY CENTER (EPIC ZOOM) IN 4KHDR | MILKY WAY GIGAPIXEL PICTURE.

https://blenderartists.org/t/high-resolution-milky-way-hdri-or-background-image-willing-to-pay/1176021/6

https://www.nasa.gov/multimedia/imagegallery/index.html

Oh finally something good. Those photo sites all want to take a chunk out of my wallet.

https://images.nasa.gov/

https://images.nasa.gov/details-PIA03653

It is not an HDRI though. Well, is that really such a problem. I could always composit it.

https://visibleearth.nasa.gov/images/57747/blue-marble-clouds

He probably got the cloud map from here.

11:40am. This is too much of a hassle. Forget outside resources. I'll just make my own. For now let me use the limbo HDRI.

The stars are too big. I'll definitely have to work on my own, but nevermind that for now. I'll procedurally generate something else later.

https://youtu.be/9Q8PwcDzb8Y?t=2277

Hmmm, do you not need the cryptomatte to get the object index. There is even a normal pass. Remember when I was looking for a way to get the camera normal. This could be it. Though it might be the world normals.

11:50am. You know what, I am an idiot. I do not need a HDRI. I can just use a plane and put an regular image on it. It is not like I am going to be rotating around the space. In that case, who cares.

Still, finding something I can use commercially without paying is a challenge. I should procedurally generate the env texture.

12pm. Sigh, it is so annoying to do the nebulae. I forgot how much time procedural texture generation can take. I absolutely must use a photo.

I did the stars, but when I rotate I get weird effects.

https://mymodernmet.com/nasa-photos-online/
NASA Lets You Download Thousands of High-Resolution Space Images for Free

I am just going to composit this thing. I am absolutely on the wrong track if I say I want to make my space procedurally.

https://images.nasa.gov/details-PIA12348

Let me go with this. This is beautiful.

Hmmm, is this really not an HDRI. It does look like one based on the distortion near the bottom.

Nope, it is definitely not an HDRI. The sides don't even mirror correctly. And and the top and bottom there is huge distortion. It does not matter. My aiming for an illustration makes things a lot easier. 3d offers a lot to a 2d artist.

I think it is finally time I get familiar with compositing.

12:20pm. What happened now is what happens when you know about something, but haven't fully ingrained it. It is really obvious that I should be compositing, but I am getting hung up on trying to get a HDRI.

https://youtu.be/a1weKAFJx3k
Demystifying Blender Cycles' Render Passes: Normal Pass

I should get familiar with passes. I am really wondering what the object index is. I'll try putting a few cubes into the scene just to see what comes out.

12:20pm. Hmmm, object index shows nothing. The normal pass shows the, well not the camera normals, but the object normals. It is pretty useless. But I should figure out how the object index works.

12:30pm. For an updated workflow I could just try using the cryptomattes, but I want to see what he does.

Yeah, this is the right way. I should just use the images from NASA as they are free, or other free photos and composit them. That is what I need to focus on. If I start getting anxious about HDRIs or even worse, try to do them procedurally I am going to be here for a long time. Rather than doing them procedurally, even just painting the nebulae on CSP would be better as I'll be able to make steady progress instead of fiddling with noise. But that would still take too much time, I'd need to dedicate at least a few days to do it.

The reason pros are fast is because they use assets, not because they can do everything on their own much quicker than the others. I need to focus on that.

Let me have breakfast here."