"10:35am. I am up. Let me chill a bit. Last night my brain finally started working again.
https://forums.odforce.net/topic/50964-how-are-variations-in-labs-building-generator-utility-supposed-to-be-used/

I asked here, but did not get an answer. Yesterday I checked out the point cloud outputs and it is good. The points have orientation and scale in them. So even if I can't pass in the geo variants directly, I can still use the base as a template for regular scattering. It is no trouble.

10:40am. Yesterday I was just overwhelmed, but today I am ready to build up the pieces one by one. I will manage it somehow.

https://www.lightnovelpub.com/novel/the-legendary-mechanic-novel-10062255/245-chapter-254

Let me sate my addiction for a bit and then I will get started. My first priority will be to finish the city building tutorials and then move on from that.

11:10am. https://www.sidefx.com/tutorials/city-building-with-osm-data/

That was fun. Let me start.

I am going to watch part 2 again and then watch 3 and 4. Then I'll also watch the other vid as well.

Oh, it is possible to disable the top and bottom ledges.

Yeah, it is coming to me now. I seem to be actually capable of paying attention today.

11:20am. I find it strange how the two ledges are split into so many faces, unlike the walls.

> So we will using this cloud point for our instancing. I guess I should be using those if I want to scatter. I could export it to something like Clarisse and make it work there.

11:40am. It is going smoothly into my brain now. Yesterday this video completely perplexed me, but now I can grasp it. It seems Unreal itself has some ability to do instancing. I'll have to tackle one of the engines, Unity or Unreal at some point, but not now. Since I an F# fan, I'll very probably go with Unity.

11:45am. Focus me, stop reading /a/, go to the next video.

Hmmmm, Houdini has a Road Tool SOP.

11:50am. What was polyframe used again? I'll figure it out when I get to it. Once I finish watching the tutorials I will actually try following them.

12:05pm. Done with part 3. It is time for part 4.

12:10pm. It is just a hunch, but I think that Unreal is capable of handling more geometry than Blender.

He will use the peak node to move points along normals. This is a great idea. I was wondering how he would move the trees inside.

12:15pm. Now he is randomly deleting points.

Hmmm, ah right. I usually just randomly distriubte them along the curve or a surface, so I never had a need to do that. The way he is doing it now will ensure that they are evenly sampled. I'll keep the technique in mind.

12:15pm. Hmmm, rather than do it like this, he would have been better off offsetting the curve first, the sampling the points and then deleting them. The reason why is because the trees can get close together at the intersections. I wonder if there is a way of deleting points that are too close to each other?

Another option is to just sample the curve and make sure that the points are relaxed enough.

Actually, that reminds me, wasn't there a relax node in there somewhere?

12:20pm. The next are trash bins. Yeah, I am wondering how he intends to place them close to the buildings. I do have an idea, but it is not good.

12:30pm. https://www.sidefx.com/tutorials/assembling-your-city-part-1/
https://www.sidefx.com/tutorials/complex-roads/

I want to watch these.

12:35pm. Mhhh, since the first vid is 30m, let me stop so I can have breakfast here. I feel like watching these videos. I do not feel a sense of perplexity like before. I should be able to absorb all of this today."