"https://youtu.be/xD0uAqi3jF8 Blender 2.8 Basics Tutorial: Scenes & Management
2:20m. I watched the above. But how would I move assets between the scenes? Nevermind that for now. I'll make it a priority to figure it out.

https://youtu.be/pr6YlbELAuA?t=18
Compositing With Multiple Scenes In Blender

Let me watch this.

2:50pm. Ok, I figured out how to move assets between scenes. I am not sure how it could be done in the past versions, but right now it is really easy to do through the asset browser.

2:55pm. Using Ctrl + L I can link an object into a new scene, after which I'd need to duplicate object properties.

https://youtu.be/r7G13NPvXuc
How to link Blender scenes, data blocks and library overrides (2020)

Let me watch this to seal the deal. After that I'll watch the Nvidia video on large scenes.

3pm. Had to do some chores. Let me resume.

https://youtu.be/r7G13NPvXuc?t=56

Yeah, this is what I figured.

View layers and scenes are really vital. I regret not learning about them as well as the asset browser earlier. It would have only taken me a day.

3:05pm. https://youtu.be/qZWGDt61Y_0
Making a Photorealistic 3d Portrait (from scratch)

This lured me. It is not what I wanted to watch, but let me check it out.

3:15pm. https://youtu.be/IJC8Xzwiulw
Top Blender Tips for Large Scale Environment Creation w/ James Tralie

What I wanted to watch was this.

https://youtu.be/IJC8Xzwiulw?t=86

Let me take a short break here.

3:45pm. That took long. Let me resume. I am starting to get into it.

https://youtu.be/IJC8Xzwiulw?t=120
> Unsplash.

I need to remember that site along with Flickr.

https://youtu.be/IJC8Xzwiulw?t=215

Ah, so this is how you activate those. Let me check them.

https://youtu.be/IJC8Xzwiulw?t=1002

Emission shader for doing volumetrics.

4:15pm. This wasn't useful, I haven't learned much I hadn't known before. But one vid is not going to kill me. It is time to get started. To start things off, how about I do a small city. Forget big and impressive. I'll increase the scale later. Right now what I should do is put the basic scene together.

4:40pm. I did a basic fence for the platform.

Right now, why don't I throw up some buildings? I'll make a small and a big one as basic cubes. Then I will scatter them around. This will be a good opportunity to get a handle on the Scatter plugin. After what I'll focus on is setting up the overall frame. I originally focused on a shot of the city from below, but given the perspective, the city did not even show up in the image at all. Instead it ended up being an UFO in the distance. This is what convinced me to ditch Clarisse - because it would be way too much of a hassle to have to deal with those big assets before the basic setup has been made.

One thing I could do is go for a very non-realistic shot. It is like how the cover has a fake sun going over the horizon while the actual this is off in the distance.

What I could do is rotate the city as if it were falling, so the camera can see it from the top. And bring it a lot closer.

I thought of having it glow from below as well as have also a bunch of shinning blocks around it, but that kind of setup would mess things up.

4:50pm. I'll go for a non-realistic scene. I'll rotate the platform as in my original vision. I thought that having a view of the earth from above from the city would be too much in the cover, but this will work fine. I'll composit the scenes together.

4:55pm. Yeah, it is decided. Two scenes rolled into one non-realistically. It will be a piece of cake to do this. Let me focus on modeling some buildings.

5:05pm. Collection instances are awkward as they don't have proper pivots.

https://blenderartists.org/t/how-do-you-control-the-origin-point-of-collection-instances/1258099

> Parent everything to an empty end put the empty into the collection. The empty will be the origin of the instance.

Hmmm, really? Will this work?

Edit: I figured it out. It turns out you can set the 3d cursor to somewhere and then set the collection origin to that. It is annoying to move the collection after that though because the origin won't update.. But it makes more sense to do it like this than join everything into one object. It does not require to apply the object mods for one.

5:45pm. How annoying. But I've done the very basic low-poly props.

I do not think I am suitable for being an artist given how long the simplest of things take me.

But I'll get the illustration done at least. I have 4 buildings with either roofs or some props on a flat plane. No windows or doors on them yet.

5:50pm. They are all organized neatly in a second scene. Oh, let me mark the buildings as assets.

Ah, whops, I forgot to set the collection offset for the buildings. Silly me.

6:10pm. The asset browser is going haywire on the platform, I can't use it at all properly on it. Maybe 1km x 1km was too much. At any rate, this is finally the time to bring in Scatter. Instead of opening a new file, it is really great that I can use a separate scene to do my testing on.

Maybe I'll open a bug report. In fact let me save a duplicate here since the scene is not intricate yet.

Let me leave that for later. I'll remember it when I see the 'buggy asset browser.blend' file in the future.

Now, let me make a 100x100 plane and I will play with Scatter 5. I actually hadn't had a chance to do that yet. It is a mystery to me as to how it works.

https://youtu.be/1eFq21yT1vE?t=5256
GENSHIN CONCERT 2021 《Melodies of an Endless Journey》

Now if I could only close this so I can watch a Scatter 5 tutorial. The concert really picks up the tempo towards the end. I am finding it hard to pause it.

6:30pm. It seems Scatter 5 is not recognizing instance collections in the browser.

6:35pm. https://youtu.be/HYkMd6FX8jM
Architectural Visualization in Blender - Part1/2

Nevermind that. Let me finally watch the tutorial for this. I am wondering how is it possible to import the biomes in the asset browser.

6:45pm. Let me have lunch here. I am actually quite into today's session because I am making concrete progress as slow as it is. Whereas in the last week, I was bashing my head against unsolvable problems.

7:55pm. Agh, who is going to continue now. I'll resume it later.

I am happy with how today went. You can't go wrong with a session where you are in the zone. I couple of months more of this and I'll be a proper 3/5."