"5:20pm. I am strugging to get this to work. I've figured out what the card is, and I've renamed the right one to eth0. I've even done the correct step of doing into the interfaces file and pointing that at eth0. I do not know why, but that is not working.
///

1. burn Ubuntu 14 server in DD mode
2. install LAMP and SSH server during install
3. sudo apt-get update & sudo apt-get upgrade
4. wget http://packages.grasehotspot.org/pool/main/g/grase-repo/grase-repo_1.8_all.deb
5. sudo dpkg -i grase-repo_1.8_all.deb
6. sudo apt-get update
7. sudo apt-get install grase-www-portal grase-conf-freeradius
8.
    remember MAC of main card using ifconfig
	sudo nano /etc/default/grub
	GRUB_CMDLINE_LINUX="net.ifnames=0 biosdevname=0"
	sudo update-grub
	reset
	sudo nano /etc/udev/rules.d/70-persistent-net.rules
	swap router names as needed
9.
	sudo nano /etc/network/interfaces
	edit to refer to eth0 instead of p1p1 or whatever
///

This time I was smart enough to write all the steps down.

But that very last step is not working for me. The internet connection only seems to work if both the relevant card in the '/etc/udev/rules.d/70-persistent-net.rules' file and in the '/etc/network/interfaces' file is named p1p1. This was supposed to be very straightforward, but now Ubuntu is making me suffer and wasting my time again.

Did I miss some step?

https://askubuntu.com/questions/489914/ubuntu-14-04-network-problem
> sudo service network-manager restart

I am sure I never tried this before.

> I modified `/etc/default/grub`, did the grub update, and sure enough the internet connection goes down. I changed the names so that the correct MAC is pointing to eth0, but that did not work. What worked is changing the main card's name from `eth0` to `p1p1`. That makes the Internet work again.

This is exactly the same problem I had before. But unlike last time, renaming it in the interfaces file is not enough to fix the issue.

5:35pm. This is really strange. I am having some inspiration. Does it work if I only set it as p1p2? I get the sense that the interfaces file is not working for whatever reason, but I need to make sure of it.

Let me try again.

6:05pm. The internet works when I set the router to any other name than eth0. This might be due to a name conflict with Grase Hotspot.

Given this, my belief that the main connection should be eth0 is shaken. Does it necessarily have to be eth0, or am I misremembering and could perhaps put it as eth1 last time as well?

Whatever, let me read LM until he gets back. Maybe it will turn out to be okay. If it does not, I'll have to continue work on this. I have no idea what I will do.

7:10pm. I took the chance to have lunch.

Today really demonstrates what the problem is with trying to cultivate multiple skills. The ones not being used simply fade from memory. If I did not keep a journal with notes, trying to trace the steps would be 10x harder. This is not a matter of effort - my memory is simply not good enough to keep all that is needed in it.

This is who knows how many time that I've done this and it is still giving me trouble. A specialist could do this in half an hour.

But today is fine. I might not have any art, but realizing what is causing the crashes and resolving it is like getting a superpower. I'll be able to rest much more easily from here on out.

7:40pm. Right now I am just waiting for dad to finish lunch, so he can come here and test the hotspot. After that it will be out of my hands. I'll also request his help in cleaning up the rig. No doubt after years, there should be a lot of dust on the CPU which should reduce the efficiency of cooling.

9:35pm. I am finally done with it. I had to reinstall it from scratch, but this time I have the proper recipe.

///

1. burn Ubuntu 14 server in DD mode
2. install LAMP and SSH server during install - optional, will be downloaded off the net automatically in case you forgot
3. sudo apt-get update & sudo apt-get upgrade - dpkg will give a weird error if it is out of date
4.
    ifconfig - remember the MAC of main card for the later steps
	sudo nano /etc/default/grub
	edit it so: GRUB_CMDLINE_LINUX="net.ifnames=0 biosdevname=0"
	sudo update-grub
	reset - this will create the rules file
	sudo nano /etc/udev/rules.d/70-persistent-net.rules
	swap router names as needed
5.
	sudo nano /etc/network/interfaces
	edit to refer to eth0 instead of p1p1 - grase needs eth0 to be the internet connection
6. wget http://packages.grasehotspot.org/pool/main/g/grase-repo/grase-repo_1.8_all.deb
7. sudo dpkg -i grase-repo_1.8_all.deb
8. sudo apt-get update
9. sudo apt-get install grase-www-portal grase-conf-freeradius

///

I had the idea of shuffling some of the steps around and that turned out to resolve the issue. I didn't remember it wrong - eth0 has to be the main connection for Grase Hotspot to work. I am really glad to be finally done with it. Every time this happens my dad pesters me about that 'command' I put in to fix things. I forgot what it was so I looked up the Google Groups for my posts on the grase group and remembered what it was. It seems uploads were very slow zero and turning GRO fixed that. That has nothing to do with today's issue.

9:50pm. I am really glad that I have that thing off my hands. It ate my entire day. But it is just as well, it is not like it started particularly well. Last night I was anxious about having to get a job due to my hardware failing, but luck is still on my side. I was fully determined to drop Heaven's Key if that happened, but it seems I will be developing my art skills for a while longer.

Whether that is good for either me or the world remains to be seen."