"11:45am. I am up. I blame getting up so late on my addiction to Legendary Mechanic. So far I am on chapter 580.
Let me chill a bit here and then I will start

https://boards.4chan.org/ic/thread/6123500/thoughts-on-this-nigga

I want to watch some of the vids in this thread.

https://youtu.be/8o6-2DwWWc4
●●の意見は、絶対に聞いちゃダメです。

I'll start with this one.

https://youtu.be/e7HvWvOM6jc
【気まぐれ添削48】プロが解説します！上達のポイント！

I do not feel like watching it all, let me also watch this for a bit. Then I'll post my work from yesterday on Twitter.

12:25pm. 15m of this is enough. I am not a 2d painter just yet, so I do not need to watch this right now.

https://www.youtube.com/watch?v=t1O7LpOTBfM
Jade Moon Upon a Sea of Clouds - Disc 1: Glazed Moon Over the Tides｜Genshin Impact

The Genshin Impact OST is 10/10.

12:30pm. https://www.flickr.com/creativecommons/

Before I post anything, let me just figure out what the license of the image I used was.

Hmmm, I think I used a copy righted background. I should just find something else. Yeah, that means redoing the render and the style transfer, but it no big deal.

No wait, where the hell does it say what the license is?

Yeah, I see it. The one I used is definitely fully copyrighted.

https://flic.kr/p/WycaKs

Let me go with this instead.

12:40pm. Let me redo the render.

///

I'll go with this as the finished image. It looks too much like an UFO instead of flying city, but it is fine. I only have a vague sense of what I want to do in my mind, and I can't exactly reverse the high level representation backwards through the layers to arrive at an image like a neural net could. In programming, such a high level representation coupled with some experience and skill is all that is needed to create a program, but in art being able to visualize details would be beneficial. I am guessing that to reach the apex skill as a human, one needs such an ability. Too bad, I do not have it, but I am not upset as even this good enough.

Two days ago I had a scare that my computer is failing due to all the render time crashed. Since then I've realized that they are happening due to my CPU overheating. Also, that is probably why the Luxcore room render was crashing. I also forgot to set the GPU in the preferences, so it was doing all the processing on the CPU. 3.1 was set properly which is why it was more stable. The GPU was taking the load off the CPU. What I've done is open the side case to let the air come in, and the temp went down by 6C. This was enough to stabilize it completely.

The original is pretty simple, this is how a sketchy style would look in 3d. Flats and simple shapes everywhere as well as lack of sufficient variation, unlike in 2d where the lines would be wild and unpolished. I thought of going for more elaborate models, but it does not matter. As long as it looks nice to look at, even a sketchy look is fine. At this point rather than focusing on improving the quality of my images, I should be focusing on the speed I am putting them out. This one took far too long to finish as it is.

Let me calculate it. I am going to ignore the 7 days I wasted messing with procedural generation and kitbashing to no result. My serious attempt of doing it all by my own hand is 3-4 days. A lot of this time was spent on getting a grasp on Scatter 5. I also got a grasp on some auxiliary stuff like Blender scenes, its asset browser and view layers. Had I been more skilled a work like this would definitely have been possible in a single day. But it is fine. I am low 3/5, not the apex of humanity. I've just about reached the level where I am starting to form my own style.

This time I was focusing on not repeating the mistake from my work on the room where I did too much detailing only to have the style transfer get rid of most of it. I did the right thing by focusing on the essentials.

From here on out, I will stop work on the backgrounds and get back into sculpting for a while. I need to create the chara models. I've established a good foundation last year, but my sculpting skills are still half baked. I've only done a female base model without any clothing, hair or coloring. I also haven't rigged it. Still, I am confident it won't take me too long to get a grasp on those aspects.

I guess I'll check out what Flycat has been up to. His vids really inspired me to get into sculpting. Digital sculpting is amazing, before I got into art, I never realized such an art form even existed. It is uniquely a 21st century invention. The style transfer that I am doing here is also unique to 21st century. In the future once more advanced NNs become affordable, prompt shaping will also enter the scene as a key skill. For that reason I do not want to invest myself too much into visual aesthetics. The image I am now is 3d style sketchy - and in the future I'll just tell the NN to fill in the blanks. This style will be so powerful that I might move to 2d at that time should I want to.

I like 3d though. It is a good tool when it is not burdening you to do detailing, texturing, or crashing on you. When I form my style, the hard parts of learning it will be behind me, and I will be effective at using it for any kind of illustration.

///

1:30pm. Let me get breakfast here. Also it turns out that hotspot rig is not done yet. The uploads are messed up. Thankfully I know where to look now. I'll post the images on Twitter after the break.

2:30pm. Let me resume. First order of business is to post the images on Twitter. Then I will have to look at that hotspot rig again. It just keeps hounding me. I hope that turning off GRO will be enough.

2:55pm. Let me tweet the long chain. This time I'll just send all the style transfered images. This is one of the benefits of not having to bother with /wip/ anymore. Twitter gets pretty pictures and that place gets the middle finger. Even if I just ignore the guy slinging shit at me, why would I put in special effort in posting to that place which would give me no recognition or benefit? They think too highly of themselves.

3pm. https://twitter.com/Ghostlike/status/1541043122582462466?s=20&t=zEORqaf6owSUmryUbu2n0Q

Here it is.

Let me commit here. What I have to do now is check out that hotspot again. Let me do it."