"10am. Let me do the chores here for a bit. I finally got up earlier.
10:25am. Done with chores. Let me chill for a bit and then I will start grinding drawings. What I am going to do here is the simplest thing ever, just trace over the model, color and then shade it.

Also I've decided to change the intro scene a bit. Rather than the foggy halo which looks like an UFO is beaming up the city, I am going to do a bloom effect. and maybe some sun beams after that. That will get me what I want. Instead of using the renderer, I'll use the compositor to make the scene properly impactful.

https://youtu.be/SN8nJXn7CqI
How to add BLOOM effect in CYCLES with ease - Blender

Let me refresh my memory of this.

https://youtu.be/SN8nJXn7CqI?t=94

He is using glare, not blur. This is the only thing I wasn't really sure of.

10:35am. Ok, I see. Let me chill a bit here and then I will start.

10:50am. Let me start. It is time to draw this. Genshin Impact is downloading in the background. And at 5pm I have an appointment with a haircutter. I'll leave that out of mind and just get cracking. Let me draw Gnosis and Luna. This should not be that difficult if I use vector lines.

10:55am. This is really strange, why is it rotating with the left mouse button all of a sudden?

11:20am. Yesterday I said: 'How difficult can it be?'

Right now, even with tracing I think it is very difficult in fact.

Let me focus on just the face. Forget the body. I'll zoom in the 3d camera so I have proper room.

Also forget the vector lines. Let me try being loose and seeing where that gets me. If I can grasp the face, I'll be able to grasp anything else. If I can't do it, I'll just go back to Blender and make some 3d models. I am sure I could do those. Maybe I'll use Genshin Impact as inspiration.

11:40am. No, forget it. I am giving up on 2d.

Take programming, my skills in it at a very high level, and the way I reason out programs is not something I'd ever be able to teach anybody else. It is like my brain has a MCTS engine built in that it uses to sample possible programs and it gradually builds up features, guiding me towards the completed product.

As I write code, there is a feedback loop between that internal process and writing. The code on the screen acts as another form of memory, and this greatly resembles sketching in art.

I am not sure if even 1 in a 1,000 people could do it at my level. And since that is the case, it is not strange for me that my brain wiring might not be suitable in other domains such as art, while others' could. It is suitable for one thing, and might not be suitable for another thing. If drawing or painting was that easy, I bet a lot of the RR writers would be doing it. I am probably not the only one who had this great entrepreneurial idea. I was just the one willing to put in effort.

11:50am. I have the high level vision, but it is really difficult convert that into an actual drawing.

This is not to say that the high level features are useless, just that process itself is difficult. I've put all my effort in the past 9 months into 3d, and there is little transfer between that and 2d. I am not going to be able to draw unless I spend a significant time practicing it.

Forget it. There are other paths I can take. I'll just just put 'blender anime head' into Youtube and move from there. I already know how to sculpt heads, and will be able to use that skill as the basis.

12pm. https://www.youtube.com/results?search_query=blender+anime+head

Isn't 2022 grand. Imagine trying to take this path 10 years ago. I'd have gotten nowhere and needed to give up.

https://www.youtube.com/watch?v=t1O7LpOTBfM
Jade Moon Upon a Sea of Clouds - Disc 1: Glazed Moon Over the Tides｜Genshin Impact

This OST is 10/10. I actually downloaded and installed the game just because of it, but I won't run it right now otherwise I'd just spend the rest of the day playing it. Right now it is time for breakfast.

After that it is the usual - a ton of tutorials. All of these are hour long so it will take me a few days to go through them, but after that I'll have my charas done and ready.

12:05pm. You know, while I am at it, why not investigate Genshin's style? And that Arcana thing which made waves? I've never watched it. I might as well properly get into the 3d anime style.

Giving up has bad connotations, but if you are going to give up, it is best to do that quickly. It is like selling a stock that is down only 10%. You might taken a loss, but you still have 90% of the capital that you can put into something else. It is better to take a 10% loss and put it into a potential winner than hold it for years and lose 50% and more.

12:10pm. Ok, breakfast."